src/requirements.txt
discord.py==2.1.0
python-dotenv==0.21.0
youtube_dl==2021.12.17
PyNaCl==1.5.0
